+++ src/m_input.h
@@ -59,6 +59,7 @@ enum
   input_weapon8,
   input_weapon9,
   input_weapontoggle,
+  input_weaponlastused,
 
   input_menu_up,
   input_menu_down,