+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## New Features
+- **Last used weapon** key, inspired by Half-Life
+
 ## Changes
 
 - **Menu background can now be enabled in all menus** instead of just in setup screens